ci: always use venv for Python
Previously, the venv was set up only when Python was installed as part
of the job, which was often not the case. But the caching always looked
at what the venv directory was supposed to be, so it turned out not to
do anything at all most of the time.